Replace SafeConstantInput with asSafe
Add a new type cast called "asSafe"
This is similar to asUnknown in that the desired return type is not important, but it asks the input to not do any type conversions that would be unsafe if the value is used unprocessed.

For example, setting a variable is something that uses an unprocessed value.
An equality operator, however, returns a processed value, and thus does not need asSafe